+++ apps/blog/src/trpc/middlewares/middleware.ts
@@ -1,0 +1,3 @@
+import { t } from "../trpc";
+
+export const middleware = t.middleware;
