+++ apps/blog/src/trpc/server.ts
@@ -9,7 +9,7 @@ import { db } from "@ashgw/db";
 
 import type { AppRouter } from "~/api/router";
 import { appRouter } from "~/api/router";
-import { createCallerFactory } from "~/trpc/trpc";
+import { createCallerFactory } from "~/trpc/root";
 import { createInnerTRPCContext, createTRPCContext } from "./context";
 import { makeQueryClient } from "./query-client";
 
@@ -17,7 +17,7 @@ const innerContext = createInnerTRPCContext({
   db,
 });
 
-const context = await createTRPCContext({
+const context = createTRPCContext({
   ...innerContext,
   req: {} as NextRequest,
   res: {} as NextResponse,
